Move test to test folder
